interface/requirements.txt
Django==2.0.3
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.7.7